pyproject.toml
[tool.poetry]
name = "pycon2022-micron"
version = "0.1.0"
description = "Continuous High Volume of Data Exporting From BigQuery"
authors = ["YUNG-CHUN LU <yungchunlu@micron.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-bigquery-storage = "^2.14.0"
google-cloud-pubsub = "^2.13.1"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
black = "^22.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
